requirements.txt
Pillow==9.4.0
pkg_resources==0.0.0
